requirements.txt
PyMuPDF==1.23.3
Flask==2.3.3
pdf2docx==0.5.6
Pillow==10.0.0
Pillow==10.0.1
python_docx==0.8.11
Flask-Mail==0.9.1